pyproject.toml
[project]
name = "deckbuilder"
version = "0.1.0"
description = "Deckbuilder RESTful API for totally-not Marvel Snap cards"
authors = [
    {name = "Luke"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pymongo (>=4.13.0,<5.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
